Added multiple stop() functions, simulated input, started WheelMgr, and started switching to radians.
Multiple start/stop cycles can be done now.
We simulated missing potentiometer and encoder cable for input to the Atlas or the Axis actuators. We integrated it with the new ConfMgr.isTesting() function; to we simulate input on the testing hub, but if not, we take the actual input.
we decided the radian switch because all of the trigonometric functions use radians, so we started switching everything to radians for performance purposes.